requirements.txt
flask==3.1.2
flasgger==0.9.7.1
requests==2.32.5
marshmallow==3.21.2
psycopg2-binary==2.9.6
celery==5.3.5
gunicorn==21.2.0
flask_cors==3.0.10
pytest==9.0.2
SQLAlchemy==2.0.16
flask-limiter==3.7.0
werkzeug==3.1.3
unidecode==1.3.8
pandas==2.3.1
PyJWT==2.10.1
xlsxwriter==3.2.5
openpyxl==3.1.5
scikit-learn==1.8.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
# tensorflow==2.20.0
gevent
